requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
yt-dlp>=2023.10.13
azure-cognitiveservices-speech>=1.33.0
requests>=2.31.0 